+++ src/lib/schema/sqlite.sql
@@ -14,8 +14,8 @@ CREATE TABLE IF NOT EXISTS srvc_event (
 CREATE INDEX IF NOT EXISTS idx_srvc_event_type ON srvc_event (type);
 CREATE INDEX IF NOT EXISTS idx_srvc_event_uri ON srvc_event (uri);
 
-CREATE INDEX IF NOT EXISTS idx_srvc_event_label_answer_document
-  ON srvc_event (data->>'$.document')
+CREATE INDEX IF NOT EXISTS idx_srvc_event_label_answer_event
+  ON srvc_event (data->>'$.event')
   WHERE type = 'label-answer';
 
 CREATE INDEX IF NOT EXISTS idx_srvc_event_label_answer_label
@@ -26,14 +26,14 @@ CREATE INDEX IF NOT EXISTS idx_srvc_event_label_answer_reviewer
   ON srvc_event (data->>'$.reviewer')
   WHERE type = 'label-answer';
 
-CREATE TRIGGER IF NOT EXISTS srvc_event_label_answer_document_constraint
+CREATE TRIGGER IF NOT EXISTS srvc_event_label_answer_event_constraint
 AFTER INSERT ON srvc_event
 WHEN NEW.type = 'label-answer'
 BEGIN
-  SELECT RAISE(ABORT, 'Missing document event for label-answer')
+  SELECT RAISE(ABORT, 'Missing event for label-answer')
   WHERE NOT EXISTS (
     SELECT 1 FROM srvc_event
-    WHERE hash = NEW.data->>'$.document' AND type = 'document'
+    WHERE hash = NEW.data->>'$.event'
   );
 END;
 
@@ -41,7 +41,7 @@ CREATE TRIGGER IF NOT EXISTS srvc_event_label_answer_label_constraint
 AFTER INSERT ON srvc_event
 WHEN NEW.type = 'label-answer'
 BEGIN
-  SELECT RAISE(ABORT, 'Missing label event for label-answer')
+  SELECT RAISE(ABORT, 'Missing label for label-answer')
   WHERE NOT EXISTS (
     SELECT 1 FROM srvc_event
     WHERE hash = NEW.data->>'$.label' AND type = 'label'